fix: fix search-form tests
