+++ src/__tests__/__mocks__/use-local-storage.ts
@@ -1,0 +1,5 @@
+import { vi } from 'vitest';
+
+export * from '@/hooks/use-local-storage';
+
+export const useLocalStorage = vi.fn();
