+++ src/__tests__/main-page.state-management.test.tsx
@@ -8,11 +8,11 @@ import { delay, http, HttpResponse } from 'msw';
 import { setupServer, SetupServerApi } from 'msw/node';
 import { expect } from 'vitest';
 
+import { testCharactersSearchArray } from '@/__tests__/test-utils/test-data';
 import { apiService } from '@/services/api-service';
 import { searchQueryStorage } from '@/services/local-storage';
-import { testCharactersSearchArray } from '@/test-utils/test-data';
 
-import { MainPage } from './main-page';
+import { MainPage } from '../pages/main/main-page';
 
 let server: SetupServerApi;
 const DELAY = 150;
